Limit azcopy memory buffer size
This sets the memory buffer for azcopy to the lowest size to prevent it
being OOM-Killed in the case where restrictive resource limits are in
place for the Kubernetes pod.

It seems that the "total available memory" for the container is
misreported, which makes azcopy try and buffer way more memory than
is allowed. Luckily, we can just set an environment variable to adjust
this buffer to something very small, manually.

The only impact of this should be potentially slower transfers, which
since we're just uploading one tarball is likely entirely fine.